requirements.txt
sklearn
pandas
numpy
io
datetime
re
mysql.connector
playwright
playwright.sync_api
dotenv
os
speedtest-cli
asyncio
Crawl4AI
requests
schedule
mysqlclient
scikit-learn
python-dotenv
openpyxl